+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.4.0] - 2025-04-22
 
 ### Added
 - Add Camera preview library which shows camera preview in Android.